+++ src/mapboxgl/services/MapService.js
@@ -28,7 +28,6 @@ import { TilesetsService } from '@supermap/iclient-common/iServer/TilesetsServic
 export class MapService extends ServiceBase {
 
     constructor(url, options) {
-      console.log(options);
         super(url, options);
     }
 
@@ -48,24 +47,6 @@ export class MapService extends ServiceBase {
             projection: me.options.projection
         });
         getMapStatusService.processAsync(callback);
-    }
-
-    /**
-     * @function  MapService.prototype.getWKT
-     * @description 获取WKT。
-     * @param {RequestCallback} callback - 回调函数。
-     */
-    getWKT(callback) {
-      var me = this;
-      var getMapStatusService = new CommonMapService(`${me.url}/prjCoordSys.wkt`, {
-          proxy: me.options.proxy,
-          withCredentials: me.options.withCredentials,
-          withoutFormatSuffix: true,
-          crossOrigin: me.options.crossOrigin,
-          headers: me.options.headers,
-          projection: me.options.projection
-      });
-      getMapStatusService.processAsync(callback);
     }
 
     /**