+++ src/maplibregl/index.js
@@ -6,4 +6,3 @@ export * from './control';
 export * from './core';
 export * from './services';
 export * from './overlay';
-export * from './mapping';