+++ src/maplibregl/namespace.js
@@ -44,8 +44,6 @@ import {
   GraphMap
 } from './overlay';
 
-import { initMap } from './mapping';
-
 import {
   CommonUtil,
   SuperMap,
@@ -93,4 +91,3 @@ maplibregl.supermap.HeatMapLayer = HeatMapLayer;
 maplibregl.supermap.GraphicLayer = GraphicLayer;
 maplibregl.supermap.GraticuleLayer = GraticuleLayer;
 maplibregl.supermap.FGBLayer = FGBLayer;
-maplibregl.supermap.initMap = initMap;