+++ src/openlayers/mapping/InitMap.js
@@ -1,5 +1,4 @@
 import { MapService } from '../services/MapService';
-import { InitMapServiceBase, isPlaneProjection } from '@supermap/iclient-common/iServer/InitMapServiceBase';
 import olMap from 'ol/Map';
 import View from 'ol/View';
 import Tile from 'ol/layer/Tile';
@@ -14,7 +13,7 @@ window.proj4 = proj4;
 /**
  * @function initMap
  * @description 根据 SuperMap iServer 服务参数，创建地图与图层。目前仅支持SuperMap iServer 地图服务，创建的图层为 ol.Tile。
- * @category BaseTypes Util
+ * @category BaseTypes Util 
  * @version 11.0.1
  * @param {number} url - 服务地址,例如: http://{ip}:{port}/iserver/services/map-world/rest/maps/World。
  * @param {Object} options - 参数。
@@ -22,17 +21,13 @@ window.proj4 = proj4;
  * @param {Object} [options.viewOptions] - 视图配置，参数设置参考 {@link https://openlayers.org/en/v6.15.1/apidoc/module-ol_View-View.html} ，未设置的情况下，默认使用 SuperMap iServer 服务参数进行设置。
  * @param {Object} [options.layerOptions] - 图层配置，参数设置参考 {@link https://openlayers.org/en/v6.15.1/apidoc/module-ol_layer_Tile-TileLayer.html} ，未设置的情况下，默认使用 SuperMap iServer 服务参数进行设置。
  * @param {Object} [options.sourceOptions] - 数据源配置，参数设置参考 {@link TileSuperMapRest} ，未设置的情况下，默认使用 SuperMap iServer 服务参数进行设置。
- * @param {string} [options.proxy] - 服务代理地址。
- * @param {boolean} [options.withCredentials=false] - 请求是否携带 cookie。
- * @param {boolean} [options.crossOrigin] - 是否允许跨域请求。
- * @param {Object} [options.headers] - 请求头。
  * @returns {Object} 实例对象，对象包括地图实例，图层实例，数据源实例。
  * @usage
  * ```
  * // 浏览器
  * <script type="text/javascript" src="{cdn}"></script>
  * <script>
- *   const initMap = {namespace}.initMap(url, { mapOptions, viewOptions, layerOptions, sourceOptions });
+ *   const initMap = {namespace}.initMap;
  *
  * </script>
  * // ES6 Import
@@ -43,25 +38,25 @@ window.proj4 = proj4;
  * */
 export function initMap(url, options = {}) {
   const { mapOptions, viewOptions, layerOptions, sourceOptions } = options;
-  const initMapService = new InitMapServiceBase(MapService, url, options);
-  return initMapService.getMapInfo(async function (serviceResult, resolve, reject) {
-    if (!serviceResult || !serviceResult.result) {
-      reject('service is not work!');
-      return;
-    }
-    let { prjCoordSys, bounds } = serviceResult.result;
-    if (!get(`EPSG:${prjCoordSys.epsgCode}`) && !isPlaneProjection(prjCoordSys.type)) {
-      const wkt = await initMapService.getWKT();
-      registerProj(prjCoordSys.epsgCode, wkt, bounds);
-    }
+  return new Promise((resolve, reject) => {
+    new MapService(url).getMapInfo(async function (serviceResult) {
+      if (!serviceResult || !serviceResult.result) {
+        reject('service is not work!');
+      }
+      let { prjCoordSys, bounds } = serviceResult.result;
+      if (!get(`EPSG:${prjCoordSys.epsgCode}`) && prjCoordSys.type !== 'PCS_NON_EARTH') {
+        const wkt = await getWkt(url);
+        registerProj(prjCoordSys.epsgCode, wkt, bounds);
+      }
 
-    let map = createMap(serviceResult.result, mapOptions, viewOptions);
-    let { layer, source } = createLayer(url, serviceResult.result, sourceOptions, layerOptions);
-    map.addLayer(layer);
-    resolve({
-      map,
-      source,
-      layer
+      let map = createMap(serviceResult.result, mapOptions, viewOptions);
+      let { layer, source } = createLayer(url, serviceResult.result, sourceOptions, layerOptions);
+      map.addLayer(layer);
+      resolve({
+        map,
+        source,
+        layer
+      });
     });
   });
 }
@@ -122,4 +117,13 @@ function createLayer(url, result, sourceOptions, layerOptions) {
     ...layerOptions
   });
   return { layer, source };
+}
+
+function getWkt(url) {
+  return new Promise((resolve) => {
+    new MapService(url).getWkt(function (res) {
+      let wkt = res.result.data;
+      resolve(wkt);
+    });
+  });
 }