+++ test/test-main-mapboxgl.js
@@ -68,4 +68,3 @@ import './mapboxgl/services/ImageCollectionServiceSpec';
 
 /*mapboxgl -- mapping*/
 import './mapboxgl/mapping/WebMapSpec.js';
-import './mapboxgl/mapping/InitMapSpec.js';